pyproject.toml
[project]
name = "hand-gesture-recognition"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "cvzone>=1.5.0",
    "langchain-core>=0.3.0",
    "langchain-ollama>=0.2.0",
    "mediapipe>=0.8.0",
    "numpy>=2.0.2",
    "opencv-python>=4.13.0.92",
    "protobuf>=3.20.0",
    "tensorflow>=2.10.0",
    "python-dotenv>=1.0.0",
]
